Rebased and somehow cleaned-up some of the general interface PR. Still a lot to clean-up I think.

Benchmark results:
```
name                                                  old time/op    new time/op    delta
pkg:sigs.k8s.io/structured-merge-diff/merge goos:linux goarch:amd64
DeducedSimple-12                                         117µs ± 3%     121µs ± 3%      +4.13%  (p=0.016 n=5+5)
DeducedNested-12                                         408µs ± 2%     390µs ± 4%      -4.37%  (p=0.008 n=5+5)
DeducedNestedAcrossVersion-12                            435µs ± 2%     439µs ± 2%        ~     (p=0.421 n=5+5)
LeafConflictAcrossVersion-12                             105µs ± 4%     115µs ± 2%     +10.15%  (p=0.008 n=5+5)
MultipleApplierRecursiveRealConversion-12               1.69ms ± 1%    1.89ms ± 2%     +11.70%  (p=0.008 n=5+5)
Operations/Pod/Create-12                                 110µs ± 4%     159µs ± 3%     +44.25%  (p=0.008 n=5+5)
Operations/Pod/Apply-12                                  306µs ± 4%     392µs ± 2%     +27.90%  (p=0.008 n=5+5)
Operations/Pod/Update-12                                 189µs ± 2%     337µs ± 3%     +78.32%  (p=0.008 n=5+5)
Operations/Pod/UpdateVersion-12                          262µs ± 3%     510µs ± 3%     +94.66%  (p=0.008 n=5+5)
Operations/Node/Create-12                                160µs ± 1%     229µs ± 2%     +43.51%  (p=0.008 n=5+5)
Operations/Node/Apply-12                                 513µs ± 4%     572µs ± 2%     +11.47%  (p=0.008 n=5+5)
Operations/Node/Update-12                                285µs ± 1%     531µs ± 1%     +86.46%  (p=0.008 n=5+5)
Operations/Node/UpdateVersion-12                         389µs ± 4%     826µs ± 3%    +112.40%  (p=0.008 n=5+5)
Operations/Endpoints/Create-12                          12.2µs ± 3%    18.9µs ± 2%     +54.47%  (p=0.008 n=5+5)
Operations/Endpoints/Apply-12                            907µs ± 1%      35µs ± 3%     -96.16%  (p=0.008 n=5+5)
Operations/Endpoints/Update-12                           290µs ± 2%    1915µs ± 1%    +560.98%  (p=0.008 n=5+5)
Operations/Endpoints/UpdateVersion-12                    556µs ± 1%    3634µs ± 1%    +553.34%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/Create-12           1.36ms ± 1%    1.73ms ± 2%     +27.47%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/Apply-12            3.42ms ± 3%    4.26ms ± 2%     +24.69%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/Update-12           1.81ms ± 2%    3.31ms ± 3%     +82.74%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/UpdateVersion-12    2.30ms ± 1%    4.76ms ± 2%    +106.83%  (p=0.008 n=5+5)
pkg:sigs.k8s.io/structured-merge-diff/typed goos:linux goarch:amd64
ConvertUnstructured/Pod/From-12                          125µs ± 3%      50µs ± 2%     -60.04%  (p=0.008 n=5+5)
ConvertUnstructured/Pod/To-12                           31.3µs ± 4%     0.0µs ± 9%    -100.00%  (p=0.008 n=5+5)
ConvertUnstructured/Node/From-12                         260µs ± 2%     107µs ± 1%     -58.87%  (p=0.008 n=5+5)
ConvertUnstructured/Node/To-12                          68.5µs ± 3%     0.0µs ± 2%    -100.00%  (p=0.008 n=5+5)
ConvertUnstructured/Endpoints/From-12                   5.58ms ± 3%    2.25ms ± 3%     -59.58%  (p=0.008 n=5+5)
ConvertUnstructured/Endpoints/To-12                     1.46ms ± 1%    0.00ms ± 6%    -100.00%  (p=0.008 n=5+5)
ConvertUnstructured/CustomResourceDefinition/From-12    1.97ms ± 3%    0.78ms ± 2%     -60.36%  (p=0.008 n=5+5)
ConvertUnstructured/CustomResourceDefinition/To-12       613µs ± 2%       0µs ± 2%    -100.00%  (p=0.008 n=5+5)

name                                                  old alloc/op   new alloc/op   delta
pkg:sigs.k8s.io/structured-merge-diff/merge goos:linux goarch:amd64
DeducedSimple-12                                        33.8kB ± 0%    28.2kB ± 0%     -16.52%  (p=0.008 n=5+5)
DeducedNested-12                                         103kB ± 0%      87kB ± 0%     -15.42%  (p=0.008 n=5+5)
DeducedNestedAcrossVersion-12                            107kB ± 0%      96kB ± 0%     -10.30%  (p=0.008 n=5+5)
LeafConflictAcrossVersion-12                            31.2kB ± 0%    28.2kB ± 0%      -9.53%  (p=0.016 n=5+4)
MultipleApplierRecursiveRealConversion-12                509kB ± 0%     513kB ± 0%      +0.89%  (p=0.008 n=5+5)
Operations/Pod/Create-12                                17.6kB ± 0%    22.6kB ± 0%     +28.35%  (p=0.008 n=5+5)
Operations/Pod/Apply-12                                 58.2kB ± 0%    57.8kB ± 0%      -0.62%  (p=0.000 n=4+5)
Operations/Pod/Update-12                                26.5kB ± 0%    46.4kB ± 0%     +75.07%  (p=0.008 n=5+5)
Operations/Pod/UpdateVersion-12                         34.4kB ± 0%    69.2kB ± 0%    +101.18%  (p=0.008 n=5+5)
Operations/Node/Create-12                               26.4kB ± 0%    33.2kB ± 0%     +25.88%  (p=0.008 n=5+5)
Operations/Node/Apply-12                                98.1kB ± 0%    83.0kB ± 0%     -15.42%  (p=0.008 n=5+5)
Operations/Node/Update-12                               37.7kB ± 0%    71.0kB ± 0%     +88.18%  (p=0.008 n=5+5)
Operations/Node/UpdateVersion-12                        48.1kB ± 0%   107.9kB ± 0%    +124.16%  (p=0.008 n=5+5)
Operations/Endpoints/Create-12                          2.86kB ± 0%    3.63kB ± 0%     +26.94%  (p=0.016 n=5+4)
Operations/Endpoints/Apply-12                           46.9kB ± 0%     6.9kB ± 0%     -85.21%  (p=0.016 n=5+4)
Operations/Endpoints/Update-12                          4.47kB ± 0%  263.77kB ± 0%   +5806.21%  (p=0.008 n=5+5)
Operations/Endpoints/UpdateVersion-12                   5.07kB ± 0%  522.83kB ± 0%  +10221.50%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/Create-12            118kB ± 0%     183kB ± 0%     +54.96%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/Apply-12             409kB ± 0%     520kB ± 0%     +27.03%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/Update-12            139kB ± 0%     402kB ± 0%    +189.47%  (p=0.008 n=5+5)
Operations/CustomResourceDefinition/UpdateVersion-12     159kB ± 0%     621kB ± 0%    +290.37%  (p=0.008 n=5+5)
pkg:sigs.k8s.io/structured-merge-diff/typed goos:linux goarch:amd64
ConvertUnstructured/Pod/From-12                         35.6kB ± 0%     6.4kB ± 0%     -81.99%  (p=0.008 n=5+5)
ConvertUnstructured/Pod/To-12                           14.4kB ± 0%     0.0kB         -100.00%  (p=0.008 n=5+5)
ConvertUnstructured/Node/From-12                        75.8kB ± 0%    14.7kB ± 0%     -80.60%  (p=0.008 n=5+5)
ConvertUnstructured/Node/To-12                          28.8kB ± 0%     0.0kB             ~     (p=0.079 n=4+5)
ConvertUnstructured/Endpoints/From-12                   2.22MB ± 0%    0.29MB ± 0%     -86.97%  (p=0.008 n=5+5)
ConvertUnstructured/Endpoints/To-12                      803kB ± 0%       0kB         -100.00%  (p=0.008 n=5+5)
ConvertUnstructured/CustomResourceDefinition/From-12     596kB ± 0%      93kB ± 0%     -84.40%  (p=0.016 n=5+4)
ConvertUnstructured/CustomResourceDefinition/To-12       307kB ± 0%       0kB         -100.00%  (p=0.008 n=5+5)
```

This is noticeably slower on some operations, faster on some others:
Still, if we look at a specifically bad case like Endpoint/Update, Conversion + Update gives:
5.58ms + 290µs + 1.46ms = 7.33ms vs 2.25ms + 1915µs + 0ms = 4.16ms. 
So that's an improvement, and we can still remove one level of conversion.