@jennybuckley 2019-11-22
The improvement from this PR would mostly be in the k8s benchmarks, right? (since we would be able to stop allocating a Value struct?)