@apelisse 2019-12-02
Updated the interface and a bit of clean-up and added documentation. I'd love to merge that since it's insanely hard to rebase.