@apelisse 2019-12-02
New benchmarks stats:
```
name                                                  old time/op    new time/op    delta
pkg:sigs.k8s.io/structured-merge-diff/typed goos:linux goarch:amd64
ConvertUnstructured/Pod/From-12                          125µs ± 2%      51µs ± 3%   -58.89%  (p=0.000 n=9+10)
ConvertUnstructured/Pod/To-12                           31.3µs ± 4%     0.0µs ± 5%  -100.00%  (p=0.000 n=10+10)
ConvertUnstructured/Node/From-12                         257µs ± 4%     118µs ± 3%   -54.19%  (p=0.000 n=9+9)
ConvertUnstructured/Node/To-12                          69.3µs ± 4%     0.0µs ± 3%  -100.00%  (p=0.000 n=10+7)
ConvertUnstructured/Endpoints/From-12                   5.46ms ± 2%    2.03ms ± 3%   -62.89%  (p=0.000 n=10+9)
ConvertUnstructured/Endpoints/To-12                     1.43ms ± 4%    0.00ms ± 7%  -100.00%  (p=0.000 n=10+10)
ConvertUnstructured/CustomResourceDefinition/From-12    1.89ms ± 5%    0.70ms ± 2%   -62.76%  (p=0.000 n=9+8)
ConvertUnstructured/CustomResourceDefinition/To-12       598µs ± 3%       0µs ± 5%  -100.00%  (p=0.000 n=10+10)
pkg:sigs.k8s.io/structured-merge-diff/merge goos:linux goarch:amd64
DeducedSimple-12                                         115µs ± 6%     121µs ± 3%    +5.09%  (p=0.000 n=10+10)
DeducedNested-12                                         408µs ± 4%     384µs ± 5%    -5.97%  (p=0.000 n=10+9)
DeducedNestedAcrossVersion-12                            429µs ± 4%     424µs ± 3%      ~     (p=0.315 n=9+10)
LeafConflictAcrossVersion-12                             103µs ± 3%     114µs ± 4%   +10.00%  (p=0.000 n=10+10)
MultipleApplierRecursiveRealConversion-12               1.71ms ± 3%    1.89ms ± 5%   +10.28%  (p=0.000 n=10+10)
Operations/Pod/Create-12                                 110µs ± 3%     157µs ± 5%   +42.33%  (p=0.000 n=10+10)
Operations/Pod/Apply-12                                  313µs ± 7%     387µs ± 5%   +23.83%  (p=0.000 n=10+10)
Operations/Pod/Update-12                                 190µs ± 2%     326µs ± 4%   +71.97%  (p=0.000 n=10+10)
Operations/Pod/UpdateVersion-12                          260µs ± 3%     489µs ± 3%   +88.04%  (p=0.000 n=10+10)
Operations/Node/Create-12                                159µs ± 2%     228µs ± 4%   +43.60%  (p=0.000 n=9+10)
Operations/Node/Apply-12                                 510µs ± 3%     565µs ± 2%   +10.76%  (p=0.000 n=8+10)
Operations/Node/Update-12                                280µs ± 4%     500µs ± 2%   +78.32%  (p=0.000 n=10+10)
Operations/Node/UpdateVersion-12                         388µs ± 4%     757µs ± 2%   +94.95%  (p=0.000 n=10+10)
Operations/Endpoints/Create-12                          12.2µs ± 3%    18.5µs ± 4%   +52.39%  (p=0.000 n=10+10)
Operations/Endpoints/Apply-12                            908µs ± 3%      34µs ± 4%   -96.22%  (p=0.000 n=10+10)
Operations/Endpoints/Update-12                           288µs ± 2%    1338µs ± 2%  +364.24%  (p=0.000 n=10+9)
Operations/Endpoints/UpdateVersion-12                    559µs ± 2%    2665µs ± 2%  +376.57%  (p=0.000 n=10+10)
Operations/CustomResourceDefinition/Create-12           1.37ms ± 2%    1.68ms ± 2%   +22.68%  (p=0.000 n=9+9)
Operations/CustomResourceDefinition/Apply-12            3.44ms ± 4%    4.12ms ± 3%   +19.67%  (p=0.000 n=10+9)
Operations/CustomResourceDefinition/Update-12           1.82ms ± 3%    3.09ms ± 2%   +70.08%  (p=0.000 n=10+9)
Operations/CustomResourceDefinition/UpdateVersion-12    2.30ms ± 2%    4.39ms ± 2%   +90.42%  (p=0.000 n=9+10)

name                                                  old alloc/op   new alloc/op   delta
pkg:sigs.k8s.io/structured-merge-diff/typed goos:linux goarch:amd64
ConvertUnstructured/Pod/From-12                         35.6kB ± 0%     4.6kB ± 0%   -87.01%  (p=0.000 n=10+10)
ConvertUnstructured/Pod/To-12                           14.4kB ± 0%     0.0kB       -100.00%  (p=0.000 n=7+10)
ConvertUnstructured/Node/From-12                        75.8kB ± 0%    11.5kB ± 0%   -84.88%  (p=0.000 n=10+10)
ConvertUnstructured/Node/To-12                          28.8kB ± 0%     0.0kB       -100.00%  (p=0.000 n=10+10)
ConvertUnstructured/Endpoints/From-12                   2.22MB ± 0%    0.18MB ± 0%   -92.02%  (p=0.000 n=10+10)
ConvertUnstructured/Endpoints/To-12                      803kB ± 0%       0kB       -100.00%  (p=0.000 n=10+10)
ConvertUnstructured/CustomResourceDefinition/From-12     596kB ± 0%      61kB ± 0%   -89.81%  (p=0.000 n=10+10)
ConvertUnstructured/CustomResourceDefinition/To-12       307kB ± 0%       0kB       -100.00%  (p=0.000 n=10+10)
pkg:sigs.k8s.io/structured-merge-diff/merge goos:linux goarch:amd64
DeducedSimple-12                                        33.8kB ± 0%    26.5kB ± 0%   -21.46%  (p=0.000 n=10+10)
DeducedNested-12                                         103kB ± 0%      83kB ± 0%   -19.91%  (p=0.000 n=10+10)
DeducedNestedAcrossVersion-12                            107kB ± 0%      89kB ± 0%   -16.10%  (p=0.000 n=10+9)
LeafConflictAcrossVersion-12                            31.2kB ± 0%    26.0kB ± 0%   -16.48%  (p=0.000 n=10+10)
MultipleApplierRecursiveRealConversion-12                509kB ± 0%     507kB ± 0%    -0.35%  (p=0.000 n=10+10)
Operations/Pod/Create-12                                17.6kB ± 0%    21.2kB ± 0%   +20.44%  (p=0.000 n=10+7)
Operations/Pod/Apply-12                                 58.2kB ± 0%    53.6kB ± 0%    -7.83%  (p=0.000 n=10+10)
Operations/Pod/Update-12                                26.5kB ± 0%    38.3kB ± 0%   +44.28%  (p=0.000 n=10+10)
Operations/Pod/UpdateVersion-12                         34.4kB ± 0%    54.3kB ± 0%   +57.77%  (p=0.000 n=10+10)
Operations/Node/Create-12                               26.4kB ± 0%    30.9kB ± 0%   +17.40%  (p=0.000 n=10+10)
Operations/Node/Apply-12                                98.1kB ± 0%    76.3kB ± 0%   -22.27%  (p=0.000 n=9+10)
Operations/Node/Update-12                               37.7kB ± 0%    55.0kB ± 0%   +45.75%  (p=0.000 n=10+10)
Operations/Node/UpdateVersion-12                        48.1kB ± 0%    78.1kB ± 0%   +62.23%  (p=0.000 n=10+10)
Operations/Endpoints/Create-12                          2.86kB ± 0%    3.44kB ± 0%   +20.29%  (p=0.000 n=10+8)
Operations/Endpoints/Apply-12                           46.9kB ± 0%     6.4kB ± 0%   -86.44%  (p=0.000 n=10+9)
Operations/Endpoints/Update-12                          4.47kB ± 0%    6.44kB ± 0%   +44.26%  (p=0.000 n=10+10)
Operations/Endpoints/UpdateVersion-12                   5.07kB ± 0%    8.44kB ± 0%   +66.56%  (p=0.000 n=10+10)
Operations/CustomResourceDefinition/Create-12            118kB ± 0%     170kB ± 0%   +43.89%  (p=0.000 n=10+10)
Operations/CustomResourceDefinition/Apply-12             409kB ± 0%     480kB ± 0%   +17.42%  (p=0.000 n=9+10)
Operations/CustomResourceDefinition/Update-12            139kB ± 0%     294kB ± 0%  +111.79%  (p=0.000 n=10+10)
Operations/CustomResourceDefinition/UpdateVersion-12     159kB ± 0%     418kB ± 0%  +162.71%  (p=0.000 n=6+10)
```