@apelisse 2019-12-02
Recomputing the statement I made before:
Still, if we look at a specifically bad case like Endpoint/Update, Conversion + Update gives:
5.46ms + 288µs + 1.43ms = 7.17ms vs 2.03ms + 1338µs + 0ms = 3.36ms (-0.8ms compared to initial test).