@apelisse 2019-12-03
For the custom resource we get:
1.89ms + 1.82ms + 598µs = 4.30ms vs 0.70ms + 3.09ms + 0µs = 3.79ms