@lavalamp 2019-12-03
Recall that we think we have a garbage problem, I wouldn't even look at the time numbers, they basically don't correspond to real life at all until we get our garbage under control.