@apelisse 2019-12-03
Let me recompute these numbers for number of allocs. They are somewhat correlated so I would expect the same order of magnitude but you're right.